fabtests: corrected flags argumet type in ft_sendmsg/ft_recvmsg functions
fixed fabtests send and recv functions to use flags argument type as
uint64_t instead of int as the underlying fi calls use uint64_t.
removed declaration of unused function ft_writemsg from shared.h

Also fixed functions calling ft_sendmsg and ft_recvmsg touse uint64_t for flags

+++ fabtests/common/shared.c
@@ -2923,7 +2923,7 @@ int ft_tx_msg(struct fid_ep *ep, fi_addr_t fi_addr, void *buf, size_t size, void
 }
 
 int ft_sendmsg(struct fid_ep *ep, fi_addr_t fi_addr,
-		void *buf, size_t size, void *ctx, int flags)
+		void *buf, size_t size, void *ctx, uint64_t flags)
 {
 	struct fi_msg msg;
 	struct fi_msg_tagged tagged_msg;
@@ -2963,7 +2963,7 @@ int ft_sendmsg(struct fid_ep *ep, fi_addr_t fi_addr,
 
 
 int ft_recvmsg(struct fid_ep *ep, fi_addr_t fi_addr, void *buf,
-	       size_t size, void *ctx, int flags)
+	       size_t size, void *ctx, uint64_t flags)
 {
 	struct fi_msg msg;
 	struct fi_msg_tagged tagged_msg;
